move report to Fit_VLE
